Change class names to match file names
