+++ src/multimodelplaystocksTask1.java
@@ -1,0 +1,99 @@
+/*
+* PURPOSE: Connect to InterSystems IRIS using XEP. This allow us run Native, JDBC and XEP from this connection
+*/
+
+import java.sql.SQLException;
+import java.sql.Statement;
+import java.io.IOException;
+import java.io.FileNotFoundException;
+import java.io.BufferedReader;
+import java.io.InputStreamReader;
+import java.util.HashMap;
+
+import com.intersystems.jdbc.IRIS;
+import com.intersystems.xep.Event;
+import com.intersystems.xep.EventPersister;
+import com.intersystems.xep.PersisterFactory;
+import com.intersystems.jdbc.IRISConnection;
+
+public class multimodelplaystocksTask1 {
+
+	public static void main(String[] args) {
+	    // Initialize map to store connection details from config.txt
+	    HashMap<String, String> map = new HashMap<String, String>();
+		try{
+			map = getConfig("config.txt");
+		}
+		catch (IOException e){
+			System.out.println(e.getMessage());
+		}
+
+		// Retrieve connection information from configuration file
+		String ip = map.get("ip");
+		int port = Integer.parseInt(map.get("port"));
+		String namespace = map.get("namespace");
+		String username = map.get("username");
+		String password = map.get("password");
+		
+		try {
+			// Connect to database using EventPersister, which is based on IRISDataSource
+	        EventPersister xepPersister = PersisterFactory.createPersister();
+
+	        // Connecting to database
+	        xepPersister.connect(ip, port, namespace, username, password);
+	        System.out.println("Connected to InterSystems IRIS via JDBC.");
+
+	        xepPersister.deleteExtent("Demo.StockInfo");   // Remove old test data
+	        xepPersister.importSchema("Demo.StockInfo");   // Import flat schema
+	       
+	        //***Initializations***
+	        // Create XEP Event for object access
+	        Event xepEvent = xepPersister.getEvent("Demo.StockInfo");
+
+	        // Create JDBC statement object for SQL and IRIS Native access
+	        Statement myStatement = xepPersister.createStatement();
+	        
+	        // Create IRIS Native object
+	        IRIS irisNative = IRIS.createIRIS((IRISConnection)xepPersister);
+
+			
+			// Close everything
+		    xepEvent.close();
+		    xepPersister.close();
+						
+		} catch (SQLException e) {
+			 System.out.println("Error creating stock listing");
+		}
+	}
+
+	// Helper method: Get connection details from config file
+	public static HashMap<String, String> getConfig(String filename) throws FileNotFoundException, IOException{
+        // Initial empty map to store connection details
+        HashMap<String, String> map = new HashMap<String, String>();
+
+        String line;
+
+        // Using Buffered Reader to read file
+        BufferedReader reader = new BufferedReader(new InputStreamReader(multiplayTask1.class.getResourceAsStream(filename)));
+
+        while ((line = reader.readLine()) != null)
+        {
+            // Remove all spaces and split line based on first colon
+            String[] parts = line.replaceAll("\\s+","").split(":", 2);
+
+            // Check if line contains enough information
+            if (parts.length >= 2)
+            {
+                String key = parts[0];
+                String value = parts[1];
+                map.put(key, value);
+            } else {
+                System.out.println("Ignoring line: " + line);
+            }
+        }
+
+        reader.close();
+
+        return map;
+    }
+}
