+++ src/multimodelplaystocksTask3.java
@@ -1,25 +1,30 @@
 /*
-* PURPOSE: Connect to InterSystems IRIS using XEP. This allow us run Native, JDBC and XEP from this connection
+* PURPOSE: Create StockInfo objects that can be manipulated
 */
 
+import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.io.IOException;
 import java.io.FileNotFoundException;
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
-import java.util.HashMap;
 
 import com.intersystems.jdbc.IRIS;
 import com.intersystems.xep.Event;
 import com.intersystems.xep.EventPersister;
 import com.intersystems.xep.PersisterFactory;
+
+import Demo.StockInfo;
+
 import com.intersystems.jdbc.IRISConnection;
 
-public class multiplayTask1 {
+public class multimodelplaystocksTask3 {
 
 	public static void main(String[] args) {
-	    // Initialize map to store connection details from config.txt
+		// Initialize map to store connection details from config.txt
 	    HashMap<String, String> map = new HashMap<String, String>();
 		try{
 			map = getConfig("config.txt");
@@ -46,7 +51,6 @@ public class multiplayTask1 {
 	        xepPersister.deleteExtent("Demo.StockInfo");   // Remove old test data
 	        xepPersister.importSchema("Demo.StockInfo");   // Import flat schema
 	       
-	        //***Initializations***
 	        // Create XEP Event for object access
 	        Event xepEvent = xepPersister.getEvent("Demo.StockInfo");
 
@@ -55,16 +59,83 @@ public class multiplayTask1 {
 	        
 	        // Create IRIS Native object
 	        IRIS irisNative = IRIS.createIRIS((IRISConnection)xepPersister);
+	        
+			boolean always = true;
+			Scanner scanner = new Scanner(System.in);
+			while (always) {
+				System.out.println("1. Retrieve all stock names using ADO.NET");
+				System.out.println("2. Generate sample founders and mission statements using XEP");
+				System.out.println("3. Populate values for founders and mission statements using Native API");
+				System.out.println("4. Quit");
+				System.out.print("What would you like to do? ");
 
-			
+				String option = scanner.next();
+				switch (option) {
+				case "1":
+					retrieveStock(myStatement);
+					break;
+				case "2":
+					generateSampleMissions(myStatement, xepEvent);
+					break;
+				case "3":
+					System.out.println("TO DO: Populate values for founders and mission statements");
+					break;
+				case "4":
+					System.out.println("Exited.");
+					always = false;
+					break;
+				default: 
+					System.out.println("Invalid option. Try again!");
+					break;
+				}	
+			}				
 			// Close everything
 		    xepEvent.close();
 		    xepPersister.close();
 						
 		} catch (SQLException e) {
-			 System.out.println("Error creating stock listing");
+			 System.out.println("Error creating stock listing: " + e.getMessage());
+		}
+	        
+	}
+
+	// Query all stock names using ADO.NET
+	public static void retrieveStock(Statement myStatement){
+		System.out.println("Generating stock info table...");
+			
+		// Get stock names (JDBC)
+		ResultSet myRS = myStatement.executeQuery("SELECT distinct name FROM demo.stock");
+					
+		while(myRS.next())
+		{
+			System.out.println(myRS.getString("name"));		
 		}
 	}
+
+	// Generate and store sample founder and mission statement using XEP
+	public static void generateSampleMissions(Statement myStatement, Event xepEvent){
+		// Get stock names (JDBC)
+		ResultSet myRS = myStatement.executeQuery("SELECT distinct name FROM demo.stock");
+											
+		// Create java objects and store to database (XEP)
+		ArrayList<StockInfo> stocksList = new ArrayList<StockInfo>();
+		while(myRS.next())
+		{
+			StockInfo stock = new StockInfo();
+			stock.name = myRS.getString("name");
+			System.out.println("Created stockinfo array.");
+			
+			//generate mission and founder names (Native API)
+			stock.founder = "test founder";
+			stock.mission = "some mission statement";
+			
+			System.out.println("Adding object with name " + stock.name + " founder " + stock.founder + " and mission " + stock.mission);
+			stocksList.add(stock);
+		}
+		StockInfo[] stocksArray = stocksList.toArray(new StockInfo[stocksList.size()]);
+		
+		xepEvent.store(stocksArray);
+	}	
 
 	// Helper method: Get connection details from config file
 	public static HashMap<String, String> getConfig(String filename) throws FileNotFoundException, IOException{
@@ -74,7 +145,7 @@ public class multiplayTask1 {
         String line;
 
         // Using Buffered Reader to read file
-        BufferedReader reader = new BufferedReader(new InputStreamReader(multiplayTask1.class.getResourceAsStream(filename)));
+        BufferedReader reader = new BufferedReader(new InputStreamReader(multiplayTask3.class.getResourceAsStream(filename)));
 
         while ((line = reader.readLine()) != null)
         {
