+++ src/multimodelplaystocksTask4.java
@@ -1,5 +1,11 @@
 /*
-* PURPOSE: Create StockInfo objects that can be manipulated
+* PURPOSE: This code shows how to use relational (JDBC), object (XEP), and native access side-by-side,
+* specifically to store stock company information.
+*
+* This last task adds functionality to populate test values by calling existing populate methods within InterSystems IRIS using the Native API.
+*   - JDBC is used to quickly retrieve all distinct stock names from the Demo.Stock table.
+*   - Native API is used to call population methods within InterSystems IRIS for founder and mission statement.
+*   - XEP is used to store these objects directly to the database, avoiding any translation back to tables.
 */
 
 import java.sql.ResultSet;
@@ -10,6 +16,7 @@ import java.util.HashMap;
 import java.io.IOException;
 import java.io.FileNotFoundException;
 import java.io.BufferedReader;
+import java.io.FileReader;
 import java.io.InputStreamReader;
 
 import com.intersystems.jdbc.IRIS;
@@ -21,7 +28,7 @@ import Demo.StockInfo;
 
 import com.intersystems.jdbc.IRISConnection;
 
-public class multiplayTask3 {
+public class multimodelplaystocksTask4 {
 
 	public static void main(String[] args) {
 		// Initialize map to store connection details from config.txt
@@ -78,7 +85,7 @@ public class multiplayTask3 {
 					generateSampleMissions(myStatement, xepEvent);
 					break;
 				case "3":
-					System.out.println("TO DO: Populate values for founders and mission statements");
+					populateMissions(myStatement, xepEvent);
 					break;
 				case "4":
 					System.out.println("Exited.");
@@ -87,8 +94,8 @@ public class multiplayTask3 {
 				default: 
 					System.out.println("Invalid option. Try again!");
 					break;
-				}	
-			}				
+				}
+			}	
 			// Close everything
 		    xepEvent.close();
 		    xepPersister.close();
@@ -125,9 +132,35 @@ public class multiplayTask3 {
 			stock.name = myRS.getString("name");
 			System.out.println("Created stockinfo array.");
 			
-			//generate mission and founder names (Native API)
+			// Generate mission and founder names
 			stock.founder = "test founder";
 			stock.mission = "some mission statement";
+			
+			System.out.println("Adding object with name " + stock.name + " founder " + stock.founder + " and mission " + stock.mission);
+			stocksList.add(stock);
+		}
+		StockInfo[] stocksArray = stocksList.toArray(new StockInfo[stocksList.size()]);
+		
+		xepEvent.store(stocksArray);
+	}
+
+	// Use Native API call population methods within InterSystems IRIS to populate values for founder and mission statement
+	public static void populateMissions(Statement myStatement, Event xepEvent){
+		// Get stock names (JDBC)
+		ResultSet myRS = myStatement.executeQuery("SELECT distinct name FROM demo.stock");
+		
+											
+		// Create java objects and store to database (XEP)
+		ArrayList<StockInfo> stocksList = new ArrayList<StockInfo>();
+		while(myRS.next())
+		{
+			StockInfo stock = new StockInfo();
+			stock.name = myRS.getString("name");
+			System.out.println("created stockinfo array.");
+			
+			// Generate mission and founder names (Native API)
+			stock.founder = irisNative.classMethodString("%PopulateUtils", "Name");
+			stock.mission = irisNative.classMethodString("%PopulateUtils", "Mission");
 			
 			System.out.println("Adding object with name " + stock.name + " founder " + stock.founder + " and mission " + stock.mission);
 			stocksList.add(stock);
@@ -145,7 +178,7 @@ public class multiplayTask3 {
         String line;
 
         // Using Buffered Reader to read file
-        BufferedReader reader = new BufferedReader(new InputStreamReader(multiplayTask3.class.getResourceAsStream(filename)));
+        BufferedReader reader = new BufferedReader(new InputStreamReader(multiplayTask1.class.getResourceAsStream(filename)));
 
         while ((line = reader.readLine()) != null)
         {
